% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Setting.R
\name{benchmarks}
\alias{benchmarks}
\title{HW1_work function}
\description{
Use R package \code{microbenchmark} to compare the performance of C functions (\code{gibbsR} and \code{vaccR}) and Cpp functions (\code{gibbsC} and \code{vaccC}).
}
\examples{
\dontrun{
x <- rdistPareto(5,2,2,FALSE)
print(x)
}
}
