% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCG.R
\name{CG}
\alias{CG}
\title{Conjugate gradient method}
\usage{
CG(A, b, x = 0, tol = 1e-05, max_iter = 1000)
}
\arguments{
\item{A}{Matrix of coefficients. It is required to be in matrix format.}

\item{b}{The right-hand side of a linear system. It is required to be in matrix format.}

\item{x}{Iterate the initial value. The default is an all-0 vector with the same dimension as the matrix A. It is required to be in array format.}

\item{tol}{Iterator convergence precision. The default value is \code{1e-5}.}

\item{max_iter}{The maximum number of iteration steps. If this deployment is exceeded and the accuracy is not reached, the iteration stops. The default value is \code{1e3}.}
}
\value{
Returns a list. The list includes the result vector \code{x}, number of iterations \code{iter_time}, the corresponding \code{error} and the flag \code{tag} of convergence or not.
}
\description{
Conjugate gradient method
}
