% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Gauss_Seidel}
\alias{Gauss_Seidel}
\title{Gauss-Seidel iterative method using Rcpp}
\usage{
Gauss_Seidel(A, b, x0, N = 10L)
}
\arguments{
\item{A}{Matrix of coefficients.}

\item{b}{The right-hand side of a linear system.}

\item{x0}{Iterate the initial value.}

\item{N}{Number of iterations.}
}
\value{
Results of iterative solutions for linear systems.
}
\description{
Gauss-Seidel iterative method using Rcpp
}
\examples{
\dontrun{
set.seed(1)
n <- 1e3
A <- sample(0:1,n^2,replace = TRUE,prob=c(0.95,0.05))
dim(A) <- c(n,n)
A <- -A + diag(rowSums(A)+1)
M_x <- rnorm(n)
b <- A \%*\% M_x
x0 <- array(rep(0,n))
x <- Gauss_Seidel(A,b,x0,1e3)
}
}
