% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multigrid.R
\name{Multigrid_iter}
\alias{Multigrid_iter}
\title{A single cycle of the multigrid method}
\usage{
Multigrid_iter(
  Ah,
  b,
  Qh,
  x0 = 0,
  presmooth = "Gauss-Seidel",
  v1 = 2,
  postsmooth = "Gauss-Seidel",
  v2 = 2,
  solver = "exact"
)
}
\arguments{
\item{Ah}{Coefficient matrix \code{A} Matrix of different scales on each layer of grid output in \code{Multigrid_Setup()}.}

\item{b}{The right-hand side of a linear system. It is required to be in matrix format.}

\item{Qh}{A projection array between different grids output in \code{Multigrid_Setup()}.}

\item{x0}{Iterate the initial value. It is required to be in array format. The default is an all-0 vector with the same dimension as the matrix A.}

\item{presmooth}{A pre-smooth iterator, optionally \code{c('Gauss-Seidel', 'Jacobi', 'exact')}. The default value is \code{'Gauss-Seidel'}.}

\item{v1}{The number of iteration steps of the pre-smooth iterator. The default value is 2.}

\item{postsmooth}{A post-smooth iterator, optionally \code{c('Gauss-Seidel', 'Jacobi', 'exact')}. The default value is \code{'Gauss-Seidel'}.}

\item{v2}{The number of iteration steps of the post-smooth iterator. The default value is 2.}

\item{solver}{The solver for the coarsest layer. The default is \code{'exact'}.}
}
\value{
Returns a list. The list includes the result vector \code{x} after a cycle and the corresponding \code{error}.
}
\description{
A single step loop of the multiple grid method.
}
