% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multigrid.R
\name{Multigrid_Setup}
\alias{Multigrid_Setup}
\title{Multiple grid building functions}
\usage{
Multigrid_Setup(
  A,
  layer = 0,
  CoarsenType = "Projection",
  InterpType = "Projection",
  pri = FALSE
)
}
\arguments{
\item{A}{Matrix of coefficients. It is required to be in matrix format.}

\item{layer}{The maximum number of layers. If the input is less than 1, the changes between different layers are designed according to the sparsity of matrix A by default. The default value is \code{0}.}

\item{CoarsenType}{The coarsening method. The default is '\code{Projection}'.}

\item{InterpType}{Interpolation method. The default is '\code{Projection}'}

\item{pri}{Whether to print the details of each layer. The default value is \code{FALSE}.}
}
\value{
Returns a list. The list includes the coefficient matrix \code{Ah}, the projection matrix \code{Qh}, the number of non-zero elements \code{nnz}, the \code{dense} for each layer, and the number of \code{layer}s.
}
\description{
Create different levels of multiple grids.
}
\examples{
\dontrun{
data(matrix_data_1e2)
attach(data)
before <- Multigrid_Setup(A,layer = 10,CoarsenType='Projection',
InterpType='Projection',pri = TRUE)
}

}
