% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multigrid.R
\name{Multigrid_Solve}
\alias{Multigrid_Solve}
\title{Multiple grid method solver}
\usage{
Multigrid_Solve(
  Ah,
  b,
  Qh,
  x0 = 0,
  presmooth = "Gauss-Seidel",
  v1 = 2,
  postsmooth = "Gauss-Seidel",
  v2 = 2,
  solver = "exact",
  tol = 1e-05,
  max_iter = 1000,
  Precond = FALSE
)
}
\arguments{
\item{Ah}{Coefficient matrix \code{A} Matrix of different scales on each layer of grid output in \code{Multigrid_Setup()}.}

\item{b}{The right-hand side of a linear system. It is required to be in matrix format.}

\item{Qh}{A projection array between different grids output in \code{Multigrid_Setup()}.}

\item{x0}{Iterate the initial value. It is required to be in array format. The default is an all-0 vector with the same dimension as the matrix A.}

\item{presmooth}{A pre-smooth iterator, optionally \code{c('Gauss-Seidel', 'Jacobi', 'exact')}. The default value is \code{'Gauss-Seidel'}.}

\item{v1}{The number of iteration steps of the pre-smooth iterator. The default value is 2.}

\item{postsmooth}{A post-smooth iterator, optionally \code{c('Gauss-Seidel', 'Jacobi', 'exact')}. The default value is \code{'Gauss-Seidel'}.}

\item{v2}{The number of iteration steps of the post-smooth iterator. The default value is 2.}

\item{solver}{The solver for the coarsest layer. The default is \code{'exact'}.}

\item{tol}{Precision of iteration results. The default value is \code{1e-5}.}

\item{max_iter}{The maximum number of iteration steps. If this deployment is exceeded and the accuracy is not reached, the iteration stops. The default value is \code{1e3}.}

\item{Precond}{Whether to use multiple grids as the preconditioner, if \code{TRUE} use the preconditioned conjugate gradient method. The default value is \code{FALSE}.}
}
\value{
Returns a list. The list includes the result vector \code{x}, number of iterations \code{iter_time}, the corresponding \code{error} and the flag \code{tag} of convergence or not.
}
\description{
Solve according to the constructed grid.
}
\examples{
\dontrun{
data(matrix_data_1e2)
attach(data)
before <- Multigrid_Setup(A,layer = 10,CoarsenType='Projection',InterpType='Projection',pri = TRUE)
x <- Multigrid_Solve(Ah=before$Ah,b,Qh=before$Qh,x0=0,presmooth='Gauss-Seidel',v1=5,
postsmooth='Gauss-Seidel',v2=5,solver='exact',tol=1e-6,max_iter=1e3,Precond=FALSE)
}

}
