% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCG.R
\name{PCG_MG}
\alias{PCG_MG}
\title{Precondition conjugate gradient method with AMG as the preconditioner}
\usage{
PCG_MG(
  A,
  b,
  x,
  tol = 1e-05,
  max_iter = 1000,
  Ah,
  Qh,
  presmooth = "Gauss-Seidel",
  v1 = 2,
  postsmooth = "Gauss-Seidel",
  v2 = 2,
  solver = "exact"
)
}
\arguments{
\item{A}{Matrix of coefficients. It is required to be in matrix format.}

\item{b}{The right-hand side of a linear system. It is required to be in matrix format.}

\item{x}{Iterate the initial value. The default is an all-0 vector with the same dimension as the matrix A. It is required to be in array format.}

\item{tol}{Iterator convergence precision. The default value is \code{1e-5}.}

\item{max_iter}{The maximum number of iteration steps. If this deployment is exceeded and the accuracy is not reached, the iteration stops. The default value is \code{1e3}.}

\item{Ah}{Coefficient matrix \code{A} Matrix of different scales on each layer of grid output in \code{Multigrid_Setup()}.}

\item{Qh}{A projection array between different grids output in \code{Multigrid_Setup()}.}

\item{presmooth}{A pre-smooth iterator, optionally \code{c('Gauss-Seidel', 'Jacobi', 'exact')}. The default value is \code{'Gauss-Seidel'}.}

\item{v1}{The number of iteration steps of the pre-smooth iterator. The default value is 2.}

\item{postsmooth}{A post-smooth iterator, optionally \code{c('Gauss-Seidel', 'Jacobi', 'exact')}. The default value is \code{'Gauss-Seidel'}.}

\item{v2}{The number of iteration steps of the post-smooth iterator. The default value is 2.}

\item{solver}{The solver for the coarsest layer. The default is \code{'exact'}.}
}
\value{
Returns a list. The list includes the result vector \code{x}, number of iterations \code{iter_time}, the corresponding \code{error} and the flag \code{tag} of convergence or not.
}
\description{
Precondition conjugate gradient method with AMG as the preconditioner
}
