% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RandomC}
\alias{RandomC}
\title{Gibbs sampling of bivariate normal variables using Rcpp}
\usage{
RandomC(x0, y0, N)
}
\arguments{
\item{x0}{Initial sampling value of x.}

\item{y0}{Initial sampling value of y.}

\item{N}{Number of samples.}
}
\value{
Sample results for binary variables.
}
\description{
Gibbs sampling of bivariate normal variables using Rcpp
}
