% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Setting.R
\name{rdistBETA}
\alias{rdistBETA}
\title{Return random numbers of Beta(alpha,beta) distributes}
\usage{
rdistBETA(n = 1000, alpha = 3, beta = 2, picture = FALSE)
}
\arguments{
\item{n}{random sample size}

\item{alpha}{parameters alpha}

\item{beta}{parameters beta}

\item{picture}{plot histogram}
}
\value{
random numbers of Beta(a,b) distributes
}
\description{
Return random numbers of Beta(alpha,beta) distributes
}
\examples{
\dontrun{
x <- rdistBETA(5,2,2,FALSE)
print(x)
}

}
