% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Setting.R
\name{rdistGammaExp}
\alias{rdistGammaExp}
\title{Generate random numbers of mixture Gamma_Exp distributes}
\usage{
rdistGammaExp(n = 1000, r = 4, beta = 2, picture = FALSE)
}
\arguments{
\item{n}{random sample size}

\item{r}{parameters r}

\item{beta}{parameters beta}

\item{picture}{plot histogram}
}
\value{
random numbers of mixture Gamma_Exp distributes
}
\description{
Generate random numbers of mixture Gamma_Exp distributes
}
\examples{
\dontrun{
x <- rdistGammaExp(5,4,2,FALSE)
print(x)
}

}
