% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restriction.R
\name{restriction}
\alias{restriction}
\title{Grid coarsening function}
\usage{
restriction(Matrix_A, n = 2, CoarsenType = "Projection")
}
\arguments{
\item{Matrix_A}{The coefficient matrix to be coarsened. It is required to be in matrix format.}

\item{n}{Degree of coarsening, that is, the number of nodes merged each time. The default is 2.}

\item{CoarsenType}{The coarsening method. The default is '\code{Projection}'.}
}
\value{
The projection matrix between the two layers.
}
\description{
Grid coarsening function
}
\examples{
\dontrun{
data(matrix_data_1e2)
attach(data)
A <- data$A
b <- data$b
Q <- restriction(A,n=2,CoarsenType='Projection')
}
}
