% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Smoother.R
\name{Smoother}
\alias{Smoother}
\title{Smooth function}
\usage{
Smoother(
  Matrix_A,
  Matrix_b,
  x0 = 0,
  N = 100,
  tol = -1,
  max_iter = 1000,
  SmoothMethod = "Gauss-Seidel"
)
}
\arguments{
\item{Matrix_A}{Matrix of coefficients. It is required to be in matrix format.}

\item{Matrix_b}{The right-hand side of a linear system. It is required to be in matrix format.}

\item{x0}{Iterate the initial value. The default is an all-0 vector with the same dimension as the matrix A. It is required to be in array format.}

\item{N}{The number of iteration steps of the smoother. If the precision \code{tol} is set, the iteration step number parameter is meaningless. The default value is 100.}

\item{tol}{Iterator convergence precision, if not set, iterate by the number of iteration steps N. The default value is -1, that is, this parameter is not set.}

\item{max_iter}{Maximum number of iteration steps. The default value is \code{1e3}.}

\item{SmoothMethod}{Method of smoothing, optionally \code{c('Gauss-Seidel', 'Jacobi', 'exact')}. The default value is \code{'Gauss-Seidel'}.}
}
\value{
Returns a list. The list includes the result vector \code{x}, number of iterations \code{iter_time}, the corresponding \code{error}, precision of convergence \code{tol} and the flag \code{tag} of convergence or not.
}
\description{
Multiple grid each step of the iterator selection, you can set the iteration step number and precision information.
}
\examples{
\dontrun{
data(matrix_data_1e2)
attach(data)
x <- Smoother(A,b,x0=0,N=10,SmoothMethod='Gauss-Seidel')
}

}
